Refactor the db namespace and do alpha2
