+++ setup.py
@@ -19,7 +19,7 @@ setup(
     # Versions should comply with PEP440.  For a discussion on single-sourcing
     # the version across setup.py and the project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version="0.0.1-alpha",
+    version="0.0.1-alpha-2",
     description="Library for XChem Django database",
     long_description=long_description,
     # The project's main homepage.
