+++ tests/test_transfer_function.py
@@ -1,6 +1,6 @@
 import unittest
 import setup_django
-from db.models import *
+from xchem_db.models import *
 from django.db.models import Q
 from functions import db_functions
 import json
