+++ xchem_db/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DbConfig(AppConfig):
-    name = 'db'
+    name = 'xchem_db'
